Correções

+++ src/main/java/com/example/aleloprojeto/companystorage/entity/Adress.java
@@ -1,26 +1,26 @@
 package com.example.aleloprojeto.companystorage.entity;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
-
 @Entity
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-
-
 public class Adress {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer adresId;
+    private Long adresId;
 
     @Column(nullable = false)
     private String street;
@@ -39,6 +39,5 @@ public class Adress {
 
     @Column(nullable = false)
     private String Complement;
-
 
 }
